# src/main/java/se/sundsvall/invoices/api/InvoicesResource.java
package se.sundsvall.invoices.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import java.nio.charset.StandardCharsets;
import org.springframework.http.ContentDisposition;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
import se.sundsvall.dept44.common.validators.annotation.MemberOf;
import se.sundsvall.dept44.common.validators.annotation.ValidMunicipalityId;
import se.sundsvall.dept44.common.validators.annotation.ValidOrganizationNumber;
import se.sundsvall.dept44.problem.Problem;
import se.sundsvall.dept44.problem.violations.ConstraintViolationProblem;
import se.sundsvall.invoices.api.model.CustomerInvoicesParameters;
import se.sundsvall.invoices.api.model.CustomerInvoicesResponse;
import se.sundsvall.invoices.api.model.InvoiceDetailsResponse;
import se.sundsvall.invoices.api.model.InvoiceOrigin;
import se.sundsvall.invoices.api.model.InvoiceType;
import se.sundsvall.invoices.api.model.InvoicesParameters;
import se.sundsvall.invoices.api.model.InvoicesResponse;
import se.sundsvall.invoices.api.model.PdfInvoice;
import se.sundsvall.invoices.service.InvoicesService;

import static java.util.Optional.ofNullable;
import static org.springframework.http.MediaType.ALL_VALUE;
import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
import static org.springframework.http.MediaType.APPLICATION_PDF_VALUE;
import static org.springframework.http.MediaType.APPLICATION_PROBLEM_JSON_VALUE;
import static org.springframework.http.ResponseEntity.ok;

@RestController
// Regex-constrain {municipalityId} to digits so non-numeric prefixes (e.g. /swagger-ui/index.html, /actuator/...) fall through to their real
// handlers instead of being routed here and failing with a 400 on path-variable conversion.
@RequestMapping("/{municipalityId:\\d+}")
@Validated
@ApiResponse(responseCode = "400", description = "Bad request", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(oneOf = {
	Problem.class, ConstraintViolationProblem.class
})))
@ApiResponse(responseCode = "500", description = "Internal Server error", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(implementation = Problem.class)))
@ApiResponse(responseCode = "502", description = "Bad Gateway", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(implementation = Problem.class)))
@Tag(name = "Invoices", description = "Service that delivers invoice information")
class InvoicesResource {

	private final InvoicesService invoicesService;

	InvoicesResource(final InvoicesService invoicesService) {
		this.invoicesService = invoicesService;
	}

	@GetMapping(value = "/{invoiceOrigin}", produces = {
		APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE
	})
	@Operation(summary = "Returns invoices matching sent in search parameters")
	@ApiResponse(responseCode = "200", description = "Successful operation", useReturnTypeSchema = true, content = @Content(mediaType = APPLICATION_JSON_VALUE))
	@ApiResponse(responseCode = "404", description = "Not found", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(implementation = Problem.class)))
	ResponseEntity<InvoicesResponse> getInvoices(
		@Parameter(name = "municipalityId", description = "Municipality ID", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
		@Parameter(name = "invoiceOrigin", description = "Invoice origin (invoices originates from either commercial or public activities)", schema = @Schema(allowableValues = {
			"COMMERCIAL", "PUBLIC_ADMINISTRATION"
		}, example = "COMMERCIAL")) @MemberOf(value = InvoiceOrigin.class, caseSensitive = false) @PathVariable final String invoiceOrigin,
		@Valid final InvoicesParameters searchParams) {

		return ok(invoicesService.getInvoices(municipalityId, invoiceOrigin, searchParams));
	}

	@GetMapping(value = "/COMMERCIAL/{organizationNumber}/{invoiceNumber}/details", produces = APPLICATION_JSON_VALUE)
	@Operation(summary = "Returns invoice-details of an invoice")
	@ApiResponse(responseCode = "200", description = "Successful operation", useReturnTypeSchema = true, content = @Content(mediaType = APPLICATION_JSON_VALUE))
	ResponseEntity<InvoiceDetailsResponse> getInvoiceDetails(
		@Parameter(name = "municipalityId", description = "Municipality ID", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
		@PathVariable @Parameter(name = "organizationNumber", description = "Organization number of invoice issuer", example = "5565272223", required = true) @ValidOrganizationNumber final String organizationNumber,
		@PathVariable @Parameter(name = "invoiceNumber", description = "Id of invoice", example = "333444", required = true) @NotBlank final String invoiceNumber) {

		return ok(InvoiceDetailsResponse.create().withDetails(invoicesService.getInvoiceDetails(municipalityId, organizationNumber, invoiceNumber)));
	}

	@Deprecated(forRemoval = true, since = "2026-05-21")
	@GetMapping(value = "/{invoiceOrigin}/{organizationNumber}/{invoiceNumber}/pdf", produces = APPLICATION_JSON_VALUE)
	@Operation(summary = "Returns invoice in pdf-format", description = "Deprecated. Use the pdf/download endpoint, which returns the raw PDF file instead of a base64-encoded JSON body.", deprecated = true)
	@ApiResponse(responseCode = "200", description = "Successful operation", useReturnTypeSchema = true, content = @Content(mediaType = APPLICATION_JSON_VALUE))
	@ApiResponse(responseCode = "404", description = "Not found", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(implementation = Problem.class)))
	ResponseEntity<PdfInvoice> getPdfInvoice(
		@Parameter(name = "municipalityId", description = "Municipality ID", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
		@PathVariable @Parameter(name = "organizationNumber", description = "Organization number of invoice issuer", example = "5565272223", required = true) @ValidOrganizationNumber final String organizationNumber,
		@PathVariable @Parameter(name = "invoiceNumber", description = "Id of invoice", example = "333444", required = true) @NotBlank final String invoiceNumber,
		@Parameter(name = "invoiceOrigin", description = "Invoice origin (invoices originates from either commercial or public activities)", schema = @Schema(allowableValues = {
			"COMMERCIAL", "PUBLIC_ADMINISTRATION"
		}, example = "COMMERCIAL")) @MemberOf(value = InvoiceOrigin.class, caseSensitive = false) @PathVariable(name = "invoiceOrigin") final String invoiceOrigin,
		@Parameter(name = "invoiceType", description = "InvoiceType filter parameter", schema = @Schema(allowableValues = {
			"INVOICE", "CREDIT_INVOICE", "START_INVOICE", "FINAL_INVOICE", "DIRECT_DEBIT", "SELF_INVOICE", "REMINDER", "CONSOLIDATED_INVOICE", "INTERNAL_INVOICE", "OFFSET_INVOICE", "UNKNOWN"
		})) @MemberOf(value = InvoiceType.class, nullable = true) @RequestParam(value = "invoiceType", required = false) final String invoiceType) {

		return ok(invoicesService.getPdfInvoice(organizationNumber, invoiceNumber, invoiceType, municipalityId));
	}

	@GetMapping(value = "/{invoiceOrigin}/{organizationNumber}/{invoiceNumber}/pdf/download", produces = ALL_VALUE)
	@Operation(summary = "Downloads an invoice as a pdf-file, or as a zip-archive when the invoice has several pdf-files")
	@ApiResponse(responseCode = "200", description = "Successful operation", content = {
		@Content(mediaType = APPLICATION_PDF_VALUE, schema = @Schema(type = "string", format = "binary")),
		@Content(mediaType = "application/zip", schema = @Schema(type = "string", format = "binary"))
	})
	@ApiResponse(responseCode = "404", description = "Not found", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(implementation = Problem.class)))
	ResponseEntity<StreamingResponseBody> downloadInvoicePdf(
		@Parameter(name = "municipalityId", description = "Municipality ID", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
		@PathVariable @Parameter(name = "organizationNumber", description = "Organization number of invoice issuer", example = "5565272223", required = true) @ValidOrganizationNumber final String organizationNumber,
		@PathVariable @Parameter(name = "invoiceNumber", description = "Id of invoice", example = "333444", required = true) @NotBlank final String invoiceNumber,
		@Parameter(name = "invoiceOrigin", description = "Invoice origin (invoices originates from either commercial or public activities)", schema = @Schema(allowableValues = {
			"COMMERCIAL", "PUBLIC_ADMINISTRATION"
		}, example = "COMMERCIAL")) @MemberOf(value = InvoiceOrigin.class, caseSensitive = false) @PathVariable(name = "invoiceOrigin") final String invoiceOrigin,
		@Parameter(name = "invoiceType", description = "InvoiceType filter parameter", schema = @Schema(allowableValues = {
			"INVOICE", "CREDIT_INVOICE", "START_INVOICE", "FINAL_INVOICE", "DIRECT_DEBIT", "SELF_INVOICE", "REMINDER", "CONSOLIDATED_INVOICE", "INTERNAL_INVOICE", "OFFSET_INVOICE", "UNKNOWN"
		})) @MemberOf(value = InvoiceType.class, nullable = true) @RequestParam(value = "invoiceType", required = false) final String invoiceType) {

		final var invoiceFile = invoicesService.downloadInvoicePdf(organizationNumber, invoiceNumber, invoiceType, municipalityId);

		return ok()
			.headers(headers -> headers.setContentDisposition(ContentDisposition.attachment()
				.filename(invoiceFile.fileName(), StandardCharsets.UTF_8)
				.build()))
			.contentType(invoiceFile.contentType())
			.body(outputStream -> outputStream.write(ofNullable(invoiceFile.content()).orElse(new byte[0])));
	}

	@GetMapping(value = "/COMMERCIAL/customers/invoices", produces = {
		APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE
	})
	@Operation(summary = "Returns invoices for one or more customers (commercial source)")
	@ApiResponse(responseCode = "200", description = "Successful operation", useReturnTypeSchema = true, content = @Content(mediaType = APPLICATION_JSON_VALUE))
	ResponseEntity<CustomerInvoicesResponse> getInvoicesForCustomer(
		@Parameter(name = "municipalityId", description = "Municipality ID", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
		@Valid final CustomerInvoicesParameters searchParams) {

		return ok(invoicesService.getInvoicesForCustomer(municipalityId, searchParams));
	}
}